AWS/aws_agentcore/my-complex-agent/requirements.txt
bedrock-agentcore
strands-agents
boto3>=1.34.0

